[Misc] Minor fixes and refactor utils code
