+++ embedchain/loaders/google_drive.py
@@ -8,7 +8,8 @@ except ImportError:
         "Google Drive requires extra dependencies. Install with `pip install embedchain[googledrive]`"
     ) from None
 
-from langchain.document_loaders import GoogleDriveLoader as Loader, UnstructuredFileIOLoader
+from langchain.document_loaders import GoogleDriveLoader as Loader
+from langchain.document_loaders import UnstructuredFileIOLoader
 
 from embedchain.helpers.json_serializable import register_deserializable
 from embedchain.loaders.base_loader import BaseLoader
