+++ embedchain/loaders/notion.py
@@ -7,7 +7,7 @@ import requests
 
 from embedchain.helpers.json_serializable import register_deserializable
 from embedchain.loaders.base_loader import BaseLoader
-from embedchain.utils import clean_string
+from embedchain.utils.misc import clean_string
 
 
 class NotionDocument:
