+++ embedchain/loaders/pdf_file.py
@@ -8,7 +8,7 @@ except ImportError:
     ) from None
 from embedchain.helpers.json_serializable import register_deserializable
 from embedchain.loaders.base_loader import BaseLoader
-from embedchain.utils import clean_string
+from embedchain.utils.misc import clean_string
 
 
 @register_deserializable
