+++ embedchain/vectordb/elasticsearch.py
@@ -11,7 +11,7 @@ except ImportError:
 
 from embedchain.config import ElasticsearchDBConfig
 from embedchain.helpers.json_serializable import register_deserializable
-from embedchain.utils import chunks
+from embedchain.utils.misc import chunks
 from embedchain.vectordb.base import BaseVectorDB
